Fix bug:
If the service has the "letters_as_pdf" permission the validation to check they have permission to send letters failed.
There is a fix for that and new tests.

The big change here is to use the default settings when creating test services, which is why so many unit tests are changed here.