Renamed 'Pre-compiled PDF' to 'Provided as PDF'
Initially we used the term 'Pre-compiled PDF' but after a time demo to
the rest of the team we decided on 'Provided as PDF'. This updates the
config to use the term and updates any file in the code base which uses
a static string to use the configuration item so that it is easier to
maintain in future and static strings don't get hidden away.

* Updated the Config with the new name
* Updated endpoint to use config item
* Updated static strings in tests to use the config item