Use today's date when persisting sorted and unsorted count
When we receive a response file from the DVLA we need to get the date
that the notifications it contains were processed to use as the
billing_day field of the daily_sorted_letter table. Previously, we get
that from converting the date in the filename to local time.

Since we always process the files on the day they were sent, we can
simply the code by using today's date as the billing_day. We also don't
need to convert the date to local time, since these files are not
received around midnight.