Add DAO function to 'delete' email reply_to addresses
Added a new DAO function which 'deletes' email reply_to addresses by
setting 'is_active' to False. This raises an error if trying to make a
default reply_to address inactive. Instead of raising an InvalidRequest
error when trying to delete a default reply_to, we created a new error
type, ValidationError. This is because InvalidRequest can't be imported
in the service_sms_sender_dao due to issues with circular imports and we
wanted to keep the error type raised consistent for different
notification types.