Add DAO function to 'delete' letter contact blocks
Added a new DAO function which 'deletes' letter contact blocks by setting
'is_active' to False. This raises an error if trying to make either a default
letter contact block inactive or if trying to make a letter contact
block which is being using as the default for a template inactive.